Переписал на specific репозитории + unitOfWork, Добавил хранилище ошибок, сделал там параметризацию, пока только в сервисах

+++ src/Core/OrderPaymentSystem.Application/DTOs/UserRole/UpdateUserRoleDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace OrderPaymentSystem.Application.DTOs.UserRole;
 
 public record UpdateUserRoleDto(
-        long FromRoleId,
-        long ToRoleId
+        int FromRoleId,
+        int ToRoleId
     );
